make `fhomed` use `urfave/cli`
