+++ cmd/fhomed/hack.go
@@ -1,0 +1,32 @@
+package main
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+
+	"github.com/bartekpacia/fhome/api"
+)
+
+// Below is a hacky workaround for myself to open my gate from my phone.
+
+func serviceListener(client *api.Client) {
+	http.HandleFunc("/gate", func(w http.ResponseWriter, r *http.Request) {
+		var result string
+		err := client.SendEvent(260, api.ValueToggle)
+		if err != nil {
+			result = fmt.Sprintf("Failed to send event: %v", err)
+			w.WriteHeader(http.StatusInternalServerError)
+		}
+
+		if result != "" {
+			log.Print(result)
+			fmt.Fprint(w, result)
+		}
+	})
+
+	err := http.ListenAndServe(":9000", nil)
+	if err != nil {
+		panic(err)
+	}
+}
